refactor: del unused, rename
