--- scripts/test-codegen.js
@@ -1,3 +1,0 @@
-var codegen = require('postman-code-generators'),
-  supportedCodegens = codegen.getLanguageList();
-console.log(supportedCodegens);
