+++ src/components/CodeGenerationSection/CodeGenerationSection.tsx
