+++ src/lib/actions/codegen.ts
@@ -1,11 +1,11 @@
 'use server';
 
-const codegen = require('postman-code-generators');
+import codeGenerator from 'postman-code-generators';
 import type { Language, PostmanRequest, Options } from 'postman-code-generators';
 
 export async function getAvailableLanguages(): Promise<Language[]> {
   try {
-    const languages: Language[] = codegen.getLanguageList();
+    const languages: Language[] = codeGenerator.getLanguageList();
     return languages;
   } catch (error) {
     console.error('[SERVER ACTION] Error in getAvailableLanguages:', error);
@@ -26,7 +26,7 @@ export async function generateCodeSnippet(
   };
 
   return new Promise((resolve, reject) => {
-    codegen.convert(
+    codeGenerator.convert(
       langKey,
       langVariant,
       requestData,
